\name{calculate_Burdell_score}
\alias{calculate_Burdell_score}
\title{Calculate Burdell Score (as per \link{https://scoremyline.com/})}
\description{
Calculates the "Burdell" score, a number between 0 and 100, as described on \link{https://scoremyline.com/}.
This function is intended for use in numerically scoring how well you followed the line.
}
\usage{
calculate_Burdell_score(gpx_data, line_data = NULL, experience = "Pro")
}
\arguments{
  \item{gpx_data}{
    A data frame of the same format as generated by the \code{\link{calculate_deviations}} function.
    Must contain columns \code{Signed_Distance} and \code{Intersection_Distance}, among others.
  }
  \item{line_data}{
    (Optional) A data frame of the same format as generated by the \code{\link{read_line_gpx}} function.
    If not provided (\code{NULL}), the line will be inferred as the straight line connecting the first and last rows of \code{gpx_data}.
  }
  \item{experience}{
    (Optional) A string from the list of "Newbie", "Amateur", or "Pro" indicating how harshly the line is scored.
    The experience level adjusts the penalty applied to deviations from the line:
    \itemize{
      \item{"Pro"}: Most strict scoring, small deviations result in larger penalties and penalties are made more frequently.
      \item{"Amateur"}: Moderate scoring.
      \item{"Newbie"}: More lenient scoring, larger deviations are allowed before penalties increase and penalties are made less frequently.
    }
    Defaults to "Pro" if \code{experience} is not provided.
  }
}
\value{
  A numerical value between 0 and 100 representing how closely you followed the straight line.
  The score is computed by considering perpendicular deviations and applying penalties based on the experience level.
}
\examples{
# Example 1: Using explicit line_data and experience of "Amateur"
gpx_data <- read_route_gpx("route.gpx")
line_data <- read_line_gpx("line.gpx")
score <- calculate_Burdell_score(gpx_data, line_data, experience = "Amateur")

# Example 2: Inferring the line from gpx_data and using default experience of "Pro"
gpx_data <- read_route_gpx("route.gpx")
score <- calculate_Burdell_score(gpx_data)
}

